Login/logout navbar. logout function in AuthContext Link to signup from loginpage
